Split RolloutManager into InferenceController and RolloutExecutor
Squashed from:
- mv
- fmt
- Import asyncio in the inference controller
- rename
- rename cls
- Move the _server property with the code that uses it
- Move the rollout lifecycle hooks with the state they touch
